category's goalInComparableUnit is changed to double as it should be

+++ src/main/java/com/example/hpjtrackerbackend/dto/request/Category.java
@@ -19,5 +19,5 @@ public class Category {
     private Long id;
     private String name;
     private String color;
-    private Integer goalInComparableUnit;
+    private Double goalInComparableUnit;
 }
